Deprecate support for triggers in Pipeline jobs
