+++ docs/Home.md
@@ -38,6 +38,8 @@ Browse the Jenkins issue tracker to see any [open issues](https://issues.jenkins
     ([JENKINS-59117](https://issues.jenkins-ci.org/browse/JENKINS-59117))
   * Fixed problem with paths on Windows
     ([JENKINS-53840](https://issues.jenkins-ci.org/browse/JENKINS-53840))
+  * Deprecated `triggers` method in `pipelineJob` context, see [Migration](Migration#migrating-to-177)
+    ([JENKINS-53775](https://issues.jenkins-ci.org/browse/JENKINS-53775))
 * 1.76 (August 22 2019)
   * Added documentation for pitfalls when using multiple Job DSL build steps in a single job
     ([JENKINS-44142](https://issues.jenkins-ci.org/browse/JENKINS-44142))
