+++ job-dsl-core/src/main/docs/examples/javaposse/jobdsl/dsl/Project/triggers.groovy
