+++ job-dsl-core/src/main/groovy/javaposse/jobdsl/dsl/jobs/WorkflowJob.groovy
@@ -4,6 +4,7 @@ import javaposse.jobdsl.dsl.ContextHelper
 import javaposse.jobdsl.dsl.DslContext
 import javaposse.jobdsl.dsl.Job
 import javaposse.jobdsl.dsl.JobManagement
+import javaposse.jobdsl.dsl.helpers.triggers.TriggerContext
 import javaposse.jobdsl.dsl.helpers.workflow.WorkflowDefinitionContext
 
 class WorkflowJob extends Job {
@@ -32,6 +33,18 @@ class WorkflowJob extends Job {
         configure { Node project ->
             Node node = methodMissing('concurrentBuild', allowConcurrentBuild)
             project / node
+        }
+    }
+
+    @Deprecated
+    void triggers(@DslContext(TriggerContext) Closure closure) {
+        TriggerContext context = new TriggerContext(jobManagement, this)
+        ContextHelper.executeInContext(closure, context)
+
+        configure { Node project ->
+            context.triggerNodes.each {
+                project / 'triggers' << it
+            }
         }
     }
 }
